Inverse Short Term Fourier Transform + Saving Functionality (CTRL-S)
Implemented an inverse short term fourier transform tool so that it is now possible to go between recorded Short Term FFTs and the time domain.

This will finally now allow me to start working on filters! The filters are applied to the fourier domain, and then ISTFT will help me translate the filtered signal into functioning audio.

Onward to practical and useful filters for voice therapy :)