Created FFTs class and implemented STFT
Audio Filters will be using FFT objects to perform the filters on.

Notes: Consider changing the name of FFTs to STFT. While the structure is technically a collection of FFTs and the data used to create the FFTs, hence, FFTs, it is much more accurately a Short Time Fourier Transform of a signal.

SpecAnalysis.STFT does require a bit of thorough testing.